requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
cycler==0.12.1
daphne==4.1.2
Django==5.0.6
django-cors-headers==4.3.1
django-mdeditor==0.1.20
djangorestframework==3.15.1
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==2.1.1
packaging==24.1
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
pynput==1.7.7
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyOpenSSL==24.1.0
pyparsing==3.1.2
pytz==2024.2
PyYAML==6.0.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
scipy==1.13.1
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.0
sympy==1.12.1
tqdm==4.66.4
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
zope.interface==7.0.3
